requirements.txt
streamlit
torch
torchvision
pillow
requests
pandas
matplotlib


